change project structure, split up code
