--- fridge-sensor/start.sh
@@ -1,5 +1,0 @@
-# expects target ip address:port as argument, example : "127.0.0.1:8080"
-./fridge-sensor $1 Pfungstaedter &
-./fridge-sensor $1 Krombacher &
-./fridge-sensor $1 Grohe &
-./fridge-sensor $1 Salz &
