+++ fridge-server/stock_head~
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="de">
+  <head>
+    <meta charset="utf-8">
+    <meta http-equiv="refresh" content="5">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <style>
+      body {font-family: "Comic Sans MS", "Comic Sans", TSCu_Comic, cursive;}
+      table, th, td {
+      border: 1px solid black;
+      border-collapse: collapse;
+      padding: 2px 5px 2px 5px;
+      }
+      td{
+      text-align:center; 
+      vertical-align:middle;
+      }
+    </style>
+    <script type="text/javascript">
+      function moveWin()
+      {  
+        window.scroll(0,10000);
+        setTimeout('moveWin();',5000);
+      }
+    </script>
+    <title>SmartFridge!</title>
+  </head>
+  <body onLoad="moveWin();">
+    <table>
+
